--- build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    tonic_build::compile_protos("proto/helloworld.proto").unwrap();
-}
