+++ simulation-tonic/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "simulation-tonic"
+version = "0.0.2-alpha.0"
+authors = ["Gardner Vickers <gardner@vickers.me>"]
+edition = "2018"
+
+[dependencies]
+bytes = "0.4.12"
+futures-preview = "0.3.0-alpha.19"
+http = "0.1"
+prost = "0.5.0"
+simulation = {path = "../simulation"}
+tower-service = "0.3.0-alpha.2"
+
+[build-dependencies]
+tonic-build = {git = "https://github.com/gardnervickers/tonic", branch = "simulation"}
+
+[dev-dependencies]
+hyper = { version = "0.13.0-alpha.4", features = ["unstable-stream", "tcp"] }
+tonic = {git = "https://github.com/gardnervickers/tonic", branch = "simulation"}
