+++ simulation-tonic/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    tonic_build::compile_protos("tests/proto/helloworld.proto").unwrap();
+}
