+++ simulation-tonic/tests/proto/helloworld.proto
