+++ simulation/examples/bad_bank.rs
