+++ simulation/src/deterministic/network/fault/latency.rs
