+++ simulation/src/deterministic/network/fault/swizzle.rs
