+++ simulation/src/deterministic/network/inner.rs
