+++ simulation/src/deterministic/network/mod.rs
