+++ simulation/src/deterministic/network/socket/fault.rs
