+++ simulation/src/deterministic/random.rs
